msm8998: Set LOCAL_PRIVATE_PLATFORM_APIS
This change sets LOCAL_SDK_VERSION for all packages where
this is possible without breaking the build, and
LOCAL_PRIVATE_PLATFORM_APIS := true otherwise.

Setting one of these two will be made required soon, and this
is a change in preparation for that. Not setting LOCAL_SDK_VERSION
makes the app implicitly depend on the bootclasspath, which is
often not required. This change effectively makes depending on
private apis opt-in rather than opt-out.

original author:
Anton Hansson <hansson@google.com>

original commit for Lineage-16.0:
https://review.lineageos.org/#/c/LineageOS/android_device_oneplus_msm8998-common/+/224137

Change-Id: Ife291970d3459f376d68fa51b456fd4ff24f99c9